fix: NPE when deleting Barber service
The program was throwing a NullPointerException (NPE) when attempting to
 delete a BarberService object that was null.

The issue has been resolved by slightly modifying the method in
BarberServiceController. Additionally, the existing test method has been
fixed, and a new test method has been added to handle cases where the
BarberService object is null.

+++ src/main/java/com/piotrzawada/BarberShopBookingSystem/Controllers/BarberServiceModelController.java
@@ -84,6 +84,11 @@ public class BarberServiceModelController {
     public ResponseEntity<Response> delete(String name) {
         Response response = new Response();
 
+        if (service.getByName(name) == null) {
+            response.setMessage("Error: This does not exist");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
+
         service.deleteByName(name);
 
         response.setMessage("This Service has been successfully deleted");
